finish getting verse and refactor
